requirements.txt
# The server
Django==1.4
# Unlike the gunicorn defined in Heroku's Django example, we're going
# to use one of the async worker classes, "gevent". Using an async worker class
# is recommended when serving traffic directly to gunicorn (which is what
# happens under the Heroku Cedar stack).
#gunicorn
#gevent
#greenlet

# Database management and migration
psycopg2
south

# Revision management
django-reversion==1.6.0

# Search
django-haystack

# Tagging
django-taggit

# Sharing
django-shorturls
git+git://github.com/statesofpop/django-cal.git#egg=django-cal

# Authentication
django-social-auth

# Async requests
djangorestframework
requests

# Template and style helpers
git+git://github.com/mjumbewu/django-bootstrapped.git@bootstrap2#egg=django-bootstrapped==1.0.3-dev
django-uni-form
django-taggit-templatetags
django-compressor
slimit

# Testing and debugging
django-nose
django-debug-toolbar

# Time
pytz
